backend/requirement.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.3
billiard==3.3.0.23
celery==3.1.18
Django==3.0.2
django-cors-headers==3.2.1
django-extensions==2.2.5
djangorestframework==3.11.0
drf-nested-routers==0.91
kombu==3.0.37
pytz==2019.3
redis==2.10.3
six==1.13.0
sqlparse==0.3.0
